11.27

+++ app/Http/Controllers/Admin/AdvertController.php
@@ -7,7 +7,6 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Http\Model\advert;
-use zgldh\QiniuStorage\QiniuStorage;
 
 
 class AdvertController extends Controller
@@ -21,17 +20,23 @@ class AdvertController extends Controller
     {
         //获取分页信息
         //获取搜索全部信息
+        // var_dump($request->all());
 
         //对数据库进行模糊查询
         $res = advert::where('user','like','%'.$request->input('search').'%')->
-        //倒叙排列
+
         orderBy('user','asc')->
         //默认搜索5条数据
         paginate($request->input('paging',5));
+        //echo "<pre>";
+        // var_dump($res);
+
+       /* $count = advert::count('id');
+        $request['count'] = $coi
+        var_dump($count);*/
 
         //将数据传递到页面中
         return view('admins/advert/index',['res'=>$res,'request'=>$request]);
-
     }
 
     /**
@@ -61,57 +66,45 @@ class AdvertController extends Controller
             'link' => 'required',
             'pic' =>'required'
         ],[
-            'user.required' => '*商品名不能为空！',
-            'link.required' => '*链接地址不能为空！',
-            'pic.required' => '*商品图片不能为空！'
+            'user.required' => '商品名不能为空！',
+            'link.required' => '链接地址不能为空！',
+            'pic.required' => '商品图片不能为空！'
         ]);
 
         //判断是否有文件上传
         if ($request->hasFile('pic')) {
 
-             //初始化七牛云
-            $disk = QiniuStorage::disk('qiniu');
-
-            //获取文件内容
-            $file = $request->file('pic');
-
             //修改名字已时间戳生成文件命
             $name = 'Advert'.rand(1111,9999).time();
 
-            //获取文件命的后缀
+            // //获取文件命的后缀
             $suffix = $request->file('pic')->getClientOriginalExtension();
-
             //移动图片到
-            // $request->file('pic')->move('./admins/Uploads', $name.'.'.$suffix);
-
-            //拼装文件名
-            $prind = '/admins/Uploads/'.$name.'.'.$suffix;
-
-            //上传到七牛云
-             $bool = $disk->put($prind,file_get_contents($file->getRealPath()));
+            $request->file('pic')->move('./admins/Uploads', $name.'.'.$suffix);
 
         }
-            //获取广告用户
-            $res['user'] = $request->input('user');
-            //获取广告链接
-            $res['link'] = $request->input('link');
-            //获取上传的文件名
-            $res['pic'] = $prind;
+
+
+            //去除不需要的参数
+            $res = $request->except('_token');
+
             //获取当前时间戳
             $res['time'] = time();
 
-            //判断是否上传到七牛云
-            if (!$bool) {
-                //如果上传失败，回到广告添加页面提示用户
-                 return redirect('/admin/advert/')->with('create','广告图片添加失败！');
+            //修改所上传文件的名称
+            $res['pic'] = '/admins/Uploads/'.$name.'.'.$suffix;
 
-            }
+            //打印form传过来的参数
+            // var_dump($res);
+            //打印图片信息
+            // dd($res['pic']);
 
             //添加至数据库
             $data = advert::insert($res);
+            // var_dump($data);
             //判断如果成功去列表页，如果失败回到当前页面
             if($data){
-                return redirect('/admin/advert')->with('create','添加广告成功！');
+                return redirect('/admin/advert/')->with('create','添加广告成功！');
             } else {
                 return back()->withInput();
             }
@@ -159,42 +152,19 @@ class AdvertController extends Controller
         //判断是否有文件上传
         if ($request->hasFile('pic')) {
 
-            //初始化七牛云
-            $disk = QiniuStorage::disk('qiniu');
-
-            //获取文件内容
-            $file = $request->file('pic');
-
             //修改名字已时间戳生成文件命
             $name = 'Advert'.rand(1111,9999).time();
 
             // //获取文件命的后缀
             $suffix = $request->file('pic')->getClientOriginalExtension();
             //移动图片到
-            // $request->file('pic')->move('./admins/Uploads', $name.'.'.$suffix);
+            $request->file('pic')->move('./admins/Uploads', $name.'.'.$suffix);
 
-            //拼装图片信息
-            $print = '/admins/Uploads/'.$name.'.'.$suffix;
+            //获取图片的信息
+            $res['pic'] = '/admins/Uploads/'.$name.'.'.$suffix;
 
-            //上传到七牛云
-            $bool = $disk->put($print,file_get_contents($file->getRealPath()));
         }
-
-            //判断是否上传到七牛云
-            if (!$bool) {
-                //如果不成功返回页面修改页面
-                return redirect('/admin/advert/')->with('create','图片上传失败！');
-            }
-
-            //去数据库获取图片信息
-            $pic = advert::where('id',$id)->value('pic');
-
-            //删除七牛云信息
-            $disk->delete($pic);
-
-            //获取图片信息
-            $res['pic'] = $print;
-            //获取当前时间戳
+           //获取当前时间戳
             $res['time'] = time();
             //获取商户名字
             $res['user'] = $request->user;
@@ -202,6 +172,10 @@ class AdvertController extends Controller
             $res['link'] = $request->link;
             //获取状态
             $res['status'] = $request->status;
+
+
+            // //打印form传过来的参数
+            // var_dump($res);
 
             //修改数据库里面的信息
             $data = advert::where('id',$id)->update($res);
@@ -225,16 +199,26 @@ class AdvertController extends Controller
     public function destroy($id)
     {
         //获取信息
-        $res = advert::where('id',$id)->value('pic');
+        $res = advert::where('id',$id)->first();
 
-        //初始化七牛云
-        $disk = QiniuStorage::disk('qiniu');
+        // dd($res['pic']);
+        // 删除图片
+        $data = unlink('.'.$res->pic);
 
-        //删除七牛云信息
-        $data = $disk->delete($res);
+        //var_dump($data)返回的结果集是布尔型
+        // var_dump($data);
 
-        //删除数据库指定id的信息
-        $info = advert::where('id',$id)->delete();
+        if ($data) {
+            //删除数据库指定id的信息
+            $info = advert::where('id',$id)->delete();
+            //如果成功返回列表页面，失败回到当前页面
+            if ($info) {
+                return redirect('/admin/advert')->with('create','删除广告成功！');
+            } else {
+                return back();
 
+
+            }
+        }
     }
 }
